setup for prediction time

+++ zairachem/setup/standardize.py
@@ -37,6 +37,8 @@ class Standardize(object):
             std_smiles += [std_smiles_dict[cid]]
         dfc[SMILES_COLUMN] = std_smiles
         dfc.to_csv(os.path.join(self.path, COMPOUNDS_FILENAME), index=False)
-        dfv = pd.read_csv(os.path.join(self.path, VALUES_FILENAME))
-        dfv = dfv[dfv[COMPOUND_IDENTIFIER_COLUMN].isin(dfc[COMPOUND_IDENTIFIER_COLUMN])]
-        dfv.to_csv(os.path.join(self.path, VALUES_FILENAME), index=False)
+        values_file = os.path.join(self.path, VALUES_FILENAME)
+        if os.path.exists(values_file):
+            dfv = pd.read_csv(values_file)
+            dfv = dfv[dfv[COMPOUND_IDENTIFIER_COLUMN].isin(dfc[COMPOUND_IDENTIFIER_COLUMN])]
+            dfv.to_csv(os.path.join(self.path, VALUES_FILENAME), index=False)
